Limited log text size to a fixed number of lines.
Added a new integer in settingsInts.xml called "max_log_lines" which controls the maximum number of lines allowed in the text log. If the number of lines goes over this, lines are removed from the top (start) of the log until the log length is equal to max_log_lines.