Removed AsynTask memory leaks.
Having a reference to the tracker scanner object for each new results posting thread resulted in the threads not being cleaned up after finishing because the scanner still exists. The reference was replaced with a 'WeakReference' container which could be used in an identical manner and allowed the threads to be cleaned up after use.