Periodic wifi scanning queue fix.
Previously, every time scanWifi() was called a new periodicUpdate_wifi was called -- this, in turn, called scanWifi(). When the gps scan was carried out, scanWifi() was also called, and so more and more periodicUpdate_wifi calls were queued.

This has now been fixed by using a flag 'wifi_scan_in_queue' that is set to true when the wifi scan is in the queue and is automatically set to false when the queued event happens. scanWifi() now only adds a new queued wifi scan when wifi_scan_in_queue is false.