bumped version to 1.2.6.4
Fixed the X509 parsing bug/exception, which was caused by attempting to use a WeakReference for the certificate used in SSL.
Set the app to maintain a portrait layout, to avoid problems with rotation.
Updated messages to the log show when a manual scan is done and on which location.
Updated the server to add a null record to the combined database when there are no wifi records received.